configs/calao_qil_a9260: fix kernel build
The calao_qil_a9260 was not specifying an explicit kernel version, so
it ended up downloading the latest kernel version, and therefore fail
to apply the patches. This commit changes the defconfig to used a
fixed kernel version, so that the patches apply properly. It fixes a
Jenkins build failure that has been occuring since quite some time.

While we're at it, also lock down the version of the kernel headers,
as noticed by Peter Korsgaard.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>