lm-sensors: don't use host ldconfig
To know whether the libiconv library is available, lm-sensors is using
the host ldconfig, which is obviously wrong in cross-compilation.

Moreover, the lm-sensors program making use of the iconv_*() API
already does it when __UCLIBC__ is *not* defined. In this case, the
iconv_*() functions are already part of the C library, so there is no
need to link against a separate library. Therefore, this patch simply
removes the libiconv handling.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: "Yann E. MORIN" <yann.morin.1998@free.fr>